+++ app/models/feedback/request.py
@@ -4,10 +4,9 @@ from pydantic import BaseModel
 
 class FeedbackRequest(BaseModel):
     user_id: str
-    pre_text: str | None = None
-    post_text: str | None = None
     pre_score: int | None = None
     post_score: int | None = None
-    subject: str
-    chapter: str
-# 단원명 (예: "JS", "React", "전체")
+    pre_text: str | None = None
+    post_text: str | None = None
+    subject: str = "frontend"
+    chapter: str = "전체"  # 단원명 (예: "JS", "React", "전체")
