+++ app/routers/feedback_router.py
@@ -1,8 +1,11 @@
+import json
+from datetime import date
+
 import openai
 from datetime import date
-from typing import List
 
-from fastapi import APIRouter
+
+from fastapi import APIRouter, HTTPException
 from fastapi.encoders import jsonable_encoder
 from starlette.responses import JSONResponse
 
@@ -14,6 +17,10 @@ from app.utils.gpt_prompt import (
     build_pre_post_comparison_prompt,
     build_post_post_comparison_prompt
 )
+from typing import List
+
+from app.services.rag_module import retrieve_similar_docs
+from app.utils.build_feedback_prompt import build_feedback_prompt
 
 router = APIRouter()
 
@@ -42,22 +49,22 @@ async def list_feedbacks(userId: str):
     return JSONResponse(status_code=200, content=jsonable_encoder(serialized))
 
 
-@router.post("/generate-feedback", summary="사용자의 피드백을 생성", description="ChatGPT를 활용해서 해당 사용자의 피드백을 생성한다.")
-async def generate_feedback(data: FeedbackRequest):
-    chapter = data.chapter
-    user_id = data.user_id
+@router.post("/generate-feedback", response_model=List[FeedbackResponse], response_model_by_alias=True, summary="지정한 사용자의 피드백을 생성", description="해당 유저의 직전 테스트 결과와 이번 테스트 결과를 활용해서 피드백을 생성한다.")
+async def generate_feedback(data: FeedbackRequest, userId: str, subject: str):
+    data.user_id = userId
+    data.subject = data.subject
 
     # 사전 평가 여부 확인
     pre_doc = await db.feedback.find_one({
-        "info.userId": user_id,
-        "info.chapter": chapter,
+        "info.userId": data.user_id,
+        "info.subject": data.subject,
         "info.type": "pre"
     })
 
     # 현재 사후 평가 횟수 확인
     post_docs = await db.feedback.find({
-        "info.userId": user_id,
-        "info.chapter": chapter,
+        "info.userId": data.user_id,
+        "info.subject": data.subject,
         "info.type": "post"
     }).sort("info.date", -1).to_list(length=2)
 
@@ -76,40 +83,75 @@ async def generate_feedback(data: FeedbackRequest):
         prompt = build_post_post_comparison_prompt(previous_post, data)
         feedback_type = "post"
 
-    # GPT 호출
+    try:
+        base_prompt = build_feedback_prompt(data)
+
+        # RAG 검색 통합
+        context_docs = retrieve_similar_docs(data.post_text or data.pre_text or "")
+        context_text = "\n".join(context_docs)
+
+        # 프롬프트 최종 구성
+        full_prompt = f"""
+[RAG 기반 유사 학습 정보]
+{context_text}
+
+[사용자 피드백 요청]
+{base_prompt}
+"""
+    except ValueError as e:
+        raise HTTPException(status_code=400, detail=str(e))
+
+    #  GPT 호출
     response = openai.ChatCompletion.create(
         model="gpt-4o",
         messages=[
             {"role": "system", "content": "당신은 학습 성장 분석가입니다."},
-            {"role": "user", "content": prompt}
+            {"role": "user", "content": full_prompt}
         ],
         temperature=0.7,
         max_tokens=800
     )
-    feedback_text = response['choices'][0]['message']['content']
 
-    # DB 저장
-    await db.feedback.insert_one({
-        "user_id": user_id,
-        "info": {
-            "userId": user_id,
+    feedback_text = response.choices[0].message.content
+
+    # JSON parse
+    try:
+        parsed = json.loads(feedback_text)
+        info = parsed["info"]
+        scores = parsed["scores"]
+        feedback = parsed["feedback"]
+    except json.JSONDecodeError:
+        info = {
+            "userId": data.user_id,
             "date": date.today().isoformat(),
-            "subject": data.subject,
-            "chapter": chapter,
-            "type": feedback_type
-        },
-        "scores": {
-            "pre": data.pre_score,
-            "post": data.post_score
-        },
+            "subject": data.subject
+        }
+        scores = {
+            **data.scores,
+        }
+        scores["total"] = sum(scores.values())
+        feedback = {
+            "strength": {},
+            "weakness": {},
+            "final": feedback_text
+        }
+
+    #  MongoDB 저장
+    await db.feedback.insert_one({
+        "user_id": data.user_id,
+        "info": info,
+        "scores": scores,
         "feedback": {
-            "text": feedback_text
+            "strength": feedback.get("strength", {}),
+            "weakness": feedback.get("weakness", {}),
+            "final": feedback.get("final", "")
         }
     })
 
-    return {
-        "user_id": user_id,
-        "chapter": chapter,
-        "feedback_type": feedback_type,
-        "feedback": feedback_text
+    return_json = {
+        "info": info,
+        "scores": scores,
+        "feedback": feedback
     }
+
+    return JSONResponse(status_code=200, content=jsonable_encoder([return_json]))
